requirements.txt
django
python-dotenv
django-ckeditor
factory_boy
